FIXED: Search: Reverse Post Order
Applied the same fix I did for the map info bubble. May need to redo with future version of WordPress (see below).


###

http://groups.google.com/group/wordpress-geo-mashup-plugin/browse_thread/thread/1831f3291c44de53

> Is there a way to reverse the order of posts in the map info window/ 
> bubble? 

I haven't provided an option for that, though I can see it might make 
a nice feature. 
The only current way I can think of would be to re-sort the query 
results in the info 
window template, which is bad practice because it could break in 
future WordPress versions, 
but might work now: 
<?php $wp_query->posts = array_reverse( $wp_query->posts ); ?> 
-dylan- 



git-svn-id: file:///Users/michael/repos-aroundtheworld/trunk@122 f54931ca-e29a-4586-8446-2c8b6a4cd0ca